Convert a color with arbitrary transparency to a fixed color

This function takes a vector of colors and an alpha level and converts the colors to fixed colors with the specified alpha level.

Usage

adjcolors(colors, alpha)

Arguments

colors

Color vectors.

alpha

Alpha level ranging from 0 to 1.

Value

The colors with the specified alpha level.